Add files via upload
I deleted the FreqScopeViews because there is a Supercollider pug that makes them practically unusable (I will explain later). On the other hand, the three types of distortion - linear, quadratic and tanh - now work using the drop-down menu! If you choose linear, the Harshness knob jumps to 1 and becomes fixed. Try it to believe it!