Various improvements
A initial window shows up when running the program. From there the user has to choose the mode, than the main window opens.
Now the Input Source Menu's items are changing according to the mode we're using (tutorial or I can do it). 
There's a play/stop button, only in the case of the tutorial mode. For the moment it's working only when stopping the sound, it's still not able to restart it.
The plug-in is starting without any sound going on (see last line of code).