Add files via upload
Finally, after changing the whole routing a bit (I also upload a file containing a very simplified version of how it was approached for better clarity), I managed to have a drop-down menu that actually changes the inputs, a button that starts and stops the data stream, and the right plots ( except for a small annoying bug on the input gain). There is still a problem in switching to the 'I can do it' mode, but we'll talk about that tomorrow. Good night ahahah